+++ e2e/app.e2e-spec.ts
@@ -36,8 +36,6 @@ describe('CoreUI template', () => {
     button.click();
     browser.sleep(sleep);
     expect(body.getAttribute('class')).not.toContain('sidebar-minimized');
-    browser.driver.manage().window().setSize(600, 800);
-    browser.sleep(1000);
   });
 
   it('should toggle `sidebar-show` body.class on `navbar-toggler` click', () => {
@@ -55,5 +53,20 @@ describe('CoreUI template', () => {
     button2.click();
     browser.sleep(sleep);
     expect(body.getAttribute('class')).not.toContain('sidebar-show');
+  });
+
+  it('should toggle `aside-menu-lg-show` body.class on `navbar-toggler` click', () => {
+    browser.manage().window().maximize();
+    browser.sleep(1000);
+    const body = page.getBody();
+    expect(body.getAttribute('class')).not.toContain('aside-menu-lg-show');
+    const button1 = page.getByCss('.navbar-toggler.d-none.d-lg-block').last();
+    button1.click();
+    browser.sleep(sleep);
+    expect(body.getAttribute('class')).toContain('aside-menu-lg-show');
+    browser.sleep(sleep);
+    button1.click();
+    browser.sleep(sleep);
+    expect(body.getAttribute('class')).not.toContain('aside-menu-lg-show');
   });
 });