+++ src/app/app.component.ts
@@ -1,13 +1,23 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, NavigationEnd } from '@angular/router';
 
+import { IconSetService } from '@coreui/icons-angular';
+import { freeSet } from '@coreui/icons';
+
 @Component({
   // tslint:disable-next-line
   selector: 'body',
-  template: '<router-outlet></router-outlet>'
+  template: '<router-outlet></router-outlet>',
+  providers: [IconSetService],
 })
 export class AppComponent implements OnInit {
-  constructor(private router: Router) { }
+  constructor(
+    private router: Router,
+    public iconSet: IconSetService
+  ) {
+    // iconSet singleton
+    iconSet.icons = { ...freeSet };
+  }
 
   ngOnInit() {
     this.router.events.subscribe((evt) => {