+++ src/app/views/icons/icons.module.ts
@@ -1,4 +1,7 @@
 import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { IconModule } from '@coreui/icons-angular';
 
 import { CoreUIIconsComponent } from './coreui-icons.component';
 import { FlagsComponent } from './flags.component';
@@ -8,7 +11,11 @@ import { SimpleLineIconsComponent } from './simple-line-icons.component';
 import { IconsRoutingModule } from './icons-routing.module';
 
 @NgModule({
-  imports: [ IconsRoutingModule ],
+  imports: [
+    CommonModule,
+    IconsRoutingModule,
+    IconModule
+  ],
   declarations: [
     CoreUIIconsComponent,
     FlagsComponent,